+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 - Modules [FLOWER, SEEDS, OFFSPRING](./docs/Flower.md)
     - New modules, pattern-driven 16-step sequencer
+- Modules [FACETS](./docs/Facets.md)
+    - New module
+- Modules [ORBIT](./docs/Orbit.md)
+    - New module, a polyphonic stereo field spreader
 
-## 1.8.0 (in development)
+## 1.8.0
 
 - Module [HIVE](./docs/Hive.md)
     - New module, 4 channel sequencer running on a 2-dimensional hexagonal grid